==> 03_create_filtered_mlr_dataset.ipynb <==
## Imports

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import warnings
warnings.filterwarnings("ignore")


df = pd.read_csv("./datasets/mlr_spearman.csv")


df.head()

X = df.drop('target',axis=1)
y = df['target']
y


Run query to filter dataset - procude a binary mask to filter out genes accoridng to filter e.g filter = df_genes['ran_imp'] > 0.0

file = 'ADA_RANKED_1113_101.csv'
# file = 'RAN_RANKED_102.csv'
filepath = f'./csvs/gene_importances/{file}'
df_genes = pd.read_csv(filepath)
filter = df_genes['ada_imp'] > 0.0
# filter = df_genes['ran_imp'] > 0.0
df_genes = df_genes[filter]
df_genes

gene_list = df_genes['gene'].values.tolist()
print(len(gene_list))
gene_list

# add baCK target column to list
gene_list.append('target')
gene_list


df_reduced = df.filter(gene_list)



df

df_reduced

output_file = f'RED_{file}'
output = f'./datasets/reduced_features_mlr_datasets/{output_file}' 
output

df.to_csv(output, index=False)
